+++ src/convert.ts
@@ -1,6 +1,6 @@
 import type { XMLBuilder, XMLParser } from 'fast-xml-parser';
-import { GPX, Point, RouteDescription, SplitMethod, Splitter, TrackPoint } from './types';
-import haversine from 'haversine';
+import { GPX, RouteDescription, SplitMethod, Splitter } from './types';
+import { calculateDistancesFromStart, getLessThanIndex } from './utils';
 
 declare global {
   const XMLParser: XMLParser;
@@ -18,88 +18,57 @@ export const split = (data: string, parts: number, method: SplitMethod): RouteDe
   const builder = new XMLBuilder(defaultArgs);
 
   const parsed = parser.parse(data);
-
-  const chunks = method === SplitMethod.POINTS ? splitPoints(parsed, parts) : splitDistance(parsed, parts);
-
-  return chunks.map((chunk) => ({ route: builder.build(chunk.route), metadata: chunk.metadata }));
-};
-
-const splitPoints: Splitter = (parsed, parts) => {
-  const distanceFromStart = calculateDistancesFromStart(parsed.gpx.trk.trkseg.trkpt);
-  const quantity = parsed.gpx.trk.trkseg.trkpt.length;
-  const chunkLength = quantity / parts;
-
-  return new Array(parts).fill(undefined).map((_, index) => {
-    const output = structuredClone(parsed);
-
-    const startIndex = Math.floor(index * chunkLength);
-    const endIndex = Math.min(Math.ceil(startIndex + chunkLength), quantity - 1);
-
-    console.log(`chunk ${index} indices: ${startIndex}-${endIndex}`);
-
-    output.gpx.trk.trkseg.trkpt = output.gpx.trk.trkseg.trkpt.slice(startIndex, endIndex - startIndex);
-    output.gpx.metadata.name = `${output.gpx.metadata.name} part ${index + 1}`;
-    output.gpx.trk.name = `${output.gpx.trk.name} part ${index + 1}`;
-    return {
-      route: output,
-      metadata: { lengthMeters: distanceFromStart[endIndex] - distanceFromStart[startIndex] },
-    };
-  });
-};
-
-const splitDistance: Splitter = (parsed, parts) => {
-  const distanceFromStart = calculateDistancesFromStart(parsed.gpx.trk.trkseg.trkpt);
-
-  const partDistance = distanceFromStart[distanceFromStart.length - 1] / parts;
+  const points = parsed.gpx.trk.trkseg.trkpt;
+  const distanceFromStart = calculateDistancesFromStart(points);
 
   console.log(`total points: ${distanceFromStart.length}`);
   console.log(`total distance: ${distanceFromStart[distanceFromStart.length - 1]}`);
-  console.log(`chunk distance: ${partDistance}`);
+  console.log(`total points: ${points.length}`);
 
-  return new Array(parts).fill(undefined).map((_, index) => {
+  const chunks = method === SplitMethod.POINTS ? splitPoints(points, parts) : splitDistance(points, parts);
+
+  return chunks.map((chunk, index) => {
     const output = structuredClone(parsed);
-
-    const startIndex = getLessThanIndex(distanceFromStart, partDistance * index);
-    const endIndex = getLessThanIndex(distanceFromStart, partDistance * (index + 1));
+    const startIndex = chunk[0];
+    const endIndex = chunk[1];
 
     console.log(`chunk ${index} indices: ${startIndex}-${endIndex}`);
 
     output.gpx.trk.trkseg.trkpt = output.gpx.trk.trkseg.trkpt.slice(startIndex, endIndex - startIndex);
     output.gpx.metadata.name = `${output.gpx.metadata.name} part ${index + 1}`;
     output.gpx.trk.name = `${output.gpx.trk.name} part ${index + 1}`;
+
     return {
-      route: output,
-      metadata: { lengthMeters: distanceFromStart[endIndex] - distanceFromStart[startIndex] },
+      route: builder.build(output),
+      metadata: {
+        lengthMeters: distanceFromStart[endIndex] - distanceFromStart[startIndex],
+      },
     };
   });
 };
 
-export const convertTrackPoint = (trackPoint: TrackPoint): Point => ({
-  latitude: parseFloat(trackPoint['@_lat']),
-  longitude: parseFloat(trackPoint['@_lon']),
-});
+const splitPoints: Splitter = (points, parts) => {
+  const distanceFromStart = calculateDistancesFromStart(points);
+  const chunkLength = points.length / parts;
 
-export const calculateDistanceBetweenPoints = (one: Point, two: Point) => haversine(one, two, { unit: 'meter' });
+  return new Array(parts).fill(undefined).map((_, index) => {
+    const startIndex = Math.floor(index * chunkLength);
+    const endIndex = Math.min(Math.ceil(startIndex + chunkLength), points.length - 1);
 
-// calculate a list of distances from the previous points, expressed in degrees
-export const calculateDistancesFromStart = (points: TrackPoint[]): number[] =>
-  points
-    .map((point) => convertTrackPoint(point))
-    .reduce((total, current, index, list) => {
-      if (index === 0) {
-        return [0];
-      }
+    return [startIndex, endIndex];
+  });
+};
 
-      const last = list[index - 1];
-      return [...total, calculateDistanceBetweenPoints(current, last) + total[index - 1]];
-    }, []);
+const splitDistance: Splitter = (points, parts) => {
+  const distanceFromStart = calculateDistancesFromStart(points);
+  const partDistance = distanceFromStart[distanceFromStart.length - 1] / parts;
 
-// get the index of a the last value below the value for an ordered list of numbers
-export const getLessThanIndex = (quantities: number[], value: number): number =>
-  quantities.reduce((found, current, index) => {
-    if (value < current && index === 0) {
-      return 0;
-    }
+  console.log(`chunk distance: ${partDistance}`);
 
-    return value < current ? found : index;
-  }, -1);
+  return new Array(parts).fill(undefined).map((_, index) => {
+    const startIndex = getLessThanIndex(distanceFromStart, partDistance * index);
+    const endIndex = getLessThanIndex(distanceFromStart, partDistance * (index + 1));
+
+    return [startIndex, endIndex];
+  });
+};
